Feature Search product

+++ app/src/main/java/com/sol/food/domain/model/product/CaloricBreakdown.kt
@@ -2,6 +2,6 @@ package com.sol.food.domain.model.product
 
 data class CaloricBreakdown(
     val percentCarbs: Double,
-    val percentFat: Int,
+    val percentFat: Double,
     val percentProtein: Double
 )
